requirements.txt
Flask==1.1.2
gunicorn==20.0.4
python-dotenv==0.14.0
requests==2.24.0
